+++ package.json
@@ -9,7 +9,8 @@
     "@testing-library/user-event": "^7.1.2",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
-    "react-scripts": "3.4.1"
+    "react-scripts": "3.4.1",
+    "react-testing-library": "4.1.2"
   },
   "scripts": {
     "start": "react-scripts start",
