+++ src/App.css
@@ -1,4 +1,4 @@
-.Border {
+.border {
   width: 500px;
   height: 500px;
   margin: 0 auto;
